src/requirements.txt
requests
retrying
cloudshell-shell-core>=3.1,<3.2
cloudshell-automation-api>=8.3,<8.4
cloudshell-tg-breaking-point>=1.0,<1.1
cloudshell-networking-devices>=2.1,<2.2
